Fix freezing due to deadlock when opening LV2 UI
This is what was happening:

`getUI()` in juce_LV2_Wrapper.cpp (called by the LV2 UI instantiate function)
locks the message thread with a `MessageManagerLock`. The locking is done by
passing a message to the message thread that causes it to block until the
`MessageManagerLock` goes out of scope (specifically, it sends a
`MessageManager::Lock::BlockingMessage`).

However, whenever messages are processed (by
`InternalRunLoop::dispatchPendingEvents()` in juce_linux_Messaging.cpp, which
then calls the callback registered by `InternalMessageQueue` in its
constructor), the lock in the `InternalRunLoop` is obtained (via a `ScopedLock`
at the top of `dispatchPendingEvents()`). This means that when the
`BlockingMessage` is processed, *the lock is held*.

Back in `getUI()`, after the message thread is locked, the UI is created via
a call to `AudioProcessor::createEditorIfNeeded()`, which calls
`createEditor()`, which is overridden by Vital's `SynthPlugin` class in
src/plugin/synth_plugin.cpp. As UI components get created, eventually
`XWindowSystem::initialiseXDisplay()` in juce_linux_XWindowSystem.cpp gets
called, which calls `LinuxEventLoop::registerFdCallback()`, which calls
`InternalRunLoop::registerFdCallback()`, which *locks the lock in
`InternalRunLoop`*. But the lock is already held by `dispatchPendingEvents()`,
and there's the deadlock.

The fix is simply to unlock the lock whenever `dispatchPendingEvents()` calls
a callback.