WORKAROUND: Fix issues when reopening LV2 UI
When using Vital as an LV2 plugin, opening the UI (`ParentUI`
specifically, rather than `ExternalUI`, which has other issues),
closing it, and then reopening it causes graphical glitches and
crashes. As a workaround, this commit changes the code to destroy
and recreate the UI every time it's shown. This increases the time
it takes to open the UI, but avoids the issues.